pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pygithub>=2.6.1",
    "pyyaml>=6.0.2",
]
